% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTTree.R
\name{appendLog}
\alias{appendLog}
\title{Append States to a log file}
\usage{
appendLog(file, record, state)
}
\arguments{
\item{file}{the file to append to}

\item{record}{named list that holds the parameters we are logging}

\item{state}{the integer number of the state}
}
\description{
Append States to a log file
}
