% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTTree.R
\name{appendTree}
\alias{appendTree}
\title{This appends a tree to a nexus file
This assumes the taxa labels are already in the file. 
The last line of the file is assumed to be "END;". It is removed 
and "END;" is added after the tree.}
\usage{
appendTree(file, tree, name = "unnamed")
}
\arguments{
\item{file}{The tree file we are appending to}

\item{tree}{The phylo tree object}

\item{name}{The name of the tree file.}
}
\description{
This appends a tree to a nexus file
This assumes the taxa labels are already in the file. 
The last line of the file is assumed to be "END;". It is removed 
and "END;" is added after the tree.
}
