% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMats.R
\name{computeMatTDist}
\alias{computeMatTDist}
\title{Build a matrix indicating for each pairs of individuals how many intermediates there are in the transmission chain}
\usage{
computeMatTDist(record, burnin = 0.5)
}
\arguments{
\item{record}{Output from inferTTree function}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}
}
\value{
Matrix of intermediates in transmission chains between pairs of hosts
}
\description{
Build a matrix indicating for each pairs of individuals how many intermediates there are in the transmission chain
}
