% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMatWIW.R
\name{computeMatWIW}
\alias{computeMatWIW}
\title{Build a matrix of probability of who infected whom from a MCMC output}
\usage{
computeMatWIW(record, burnin = 0.5)
}
\arguments{
\item{record}{Output from inferTTree function}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}
}
\description{
Build a matrix of probability of who infected whom from a MCMC output
}

