% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{convertToCoda}
\alias{convertToCoda}
\title{Convert to coda mcmc format}
\usage{
convertToCoda(record, burnin = 0.5)
}
\arguments{
\item{record}{Output from inferTTree function}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}
}
\value{
Object of class mcmc from coda package
}
\description{
Convert to coda mcmc format
}
