% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dateLastSample}
\alias{dateLastSample}
\title{Return the date of last sample from a ttree or ctree or ptree}
\usage{
dateLastSample(x)
}
\arguments{
\item{x}{A transmission tree or colored tree or phylogenetic tree}
}
\value{
date of the last sample
}
\description{
Return the date of last sample from a ttree or ctree or ptree
}
