% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiPenTTree.R
\name{epiPenTTree}
\alias{epiPenTTree}
\title{Penalize a transmission tree based on epidemiological data}
\usage{
epiPenTTree(ttree, epiData, penaltyInfo = FALSE)
}
\arguments{
\item{ttree}{Transmission tree}

\item{epiData}{A named list of two data.frames: \cr \cr 
\strong{case.info} with row.names matching the case names and columns: 
\describe{
  \item{\dQuote{start}}{the start of the probable exposure for this case}
  \item{\dQuote{end}}{the end of the probable exposure for this case}
  \item{\dQuote{location}}{a string name of the location for this case}
} 
\strong{contact.info} with columns: 
\describe{
  \item{\dQuote{case.A}}{the name of the first case in this probable transmission pair}
  \item{\dQuote{case.B}}{the name of the second case in this probable transmission pair}
  \item{\dQuote{start}}{the start of the probable contact for the pair of cases}
  \item{\dQuote{end}}{the end of the probable contact for the pair of cases}
}}

\item{penaltyInfo}{Whether or not to return information about which penalties are broken for the given transmission tree}
}
\value{
Either a list of the numeric penalties or a list of the numeric penalties and data.frame of information about which events 
caused the penalties
}
\description{
Penalize a transmission tree based on epidemiological data
}
\author{
Matthew Gould
}
