% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{extractCTree}
\alias{extractCTree}
\title{Return the combined tree corresponding to a given iteration of the TransPhylo results}
\usage{
extractCTree(res, iteration)
}
\arguments{
\item{res}{Output from inferTTree command}

\item{iteration}{Number of the iteration to be extracted}
}
\value{
The colored tree at the specified iteeatino
}
\description{
Return the combined tree corresponding to a given iteration of the TransPhylo results
}
