% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPTree.R
\name{extractPTree}
\alias{extractPTree}
\title{Extracts phylogenetic tree from a combined phylogenetic/transmission tree}
\usage{
extractPTree(ctree)
}
\arguments{
\item{ctree}{Combined tree}
}
\value{
phylogenetic tree
}
\description{
Extracts phylogenetic tree from a combined phylogenetic/transmission tree
}
\examples{
extractPTree(simulateOutbreak())
}
