% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTTree.R
\name{extractTTree}
\alias{extractTTree}
\title{Extracts transmission tree from a combined phylogenetic/transmission tree}
\usage{
extractTTree(ctree)
}
\arguments{
\item{ctree}{Combined tree}
}
\value{
transmission tree
}
\description{
Extracts transmission tree from a combined phylogenetic/transmission tree
}
\examples{
extractTTree(simulateOutbreak())
}
