% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{getInfectionTimeDist}
\alias{getInfectionTimeDist}
\title{Extract and return distribution of infection time of a given sampled case}
\usage{
getInfectionTimeDist(record, burnin = 0.5, k, show.plot = F)
}
\arguments{
\item{record}{MCMC output produced by inferTTree}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}

\item{k}{Case whose posterior infection times are to be extracted. Either an integer or a string matching one of the case names in the data}

\item{show.plot}{Show a barplot of the distribution}
}
\value{
A vector of posterior infection times for case k
}
\description{
Extract and return distribution of infection time of a given sampled case
}
