% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{getOffspringDist}
\alias{getOffspringDist}
\title{Extract and return offspring distribution of given sampled case(s)}
\usage{
getOffspringDist(record, burnin = 0.5, k, show.plot = F)
}
\arguments{
\item{record}{MCMC output produced by inferTTree}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}

\item{k}{Case(s) whose offspring distribution are to be extracted. Either a string matching one of the case names in the data, or a vector of such strings}

\item{show.plot}{Show a barplot of the distribution}
}
\value{
Posterior offspring distribution for the case(s) in k. If length(k)==1 then a vector is returned, otherwise a matrix
}
\description{
Extract and return offspring distribution of given sampled case(s)
}
