% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{getSamplingTimeDist}
\alias{getSamplingTimeDist}
\title{Extract and return realised sampling time distribution}
\usage{
getSamplingTimeDist(record, burnin = 0.5, maxi = 2, numBins = 20,
  show.plot = F)
}
\arguments{
\item{record}{MCMC output produced by inferTTree}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}

\item{maxi}{Maximum generation time to consider}

\item{numBins}{Number of time bins to compute and display distribution}

\item{show.plot}{Show a barplot of the distribution}
}
\value{
Vector of times between becoming infected and becoming sampled in the posterior
}
\description{
Extract and return realised sampling time distribution
}
