% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTTree.R
\name{inferTTree}
\alias{inferTTree}
\title{Infer transmission tree given a phylogenetic tree}
\usage{
inferTTree(
  ptree,
  w.shape = 2,
  w.scale = 1,
  ws.shape = w.shape,
  ws.scale = w.scale,
  w.mu,
  w.sigma,
  ws.mu,
  ws.sigma,
  mcmcIterations = 1000,
  thinning = 1,
  startNeg = 100/365,
  R0,
  startOff.r = 1,
  startOff.p = 0.5,
  startPi = 0.5,
  updateNeg = TRUE,
  updateOff.r = TRUE,
  updateOff.p = FALSE,
  updatePi = TRUE,
  startCTree = NA,
  updateTTree = TRUE,
  optiStart = TRUE,
  dateT = Inf,
  epiData,
  penalize = TRUE,
  trackPenalty = FALSE,
  exposureP = 1,
  contactP = 1,
  locationP = 1
)
}
\arguments{
\item{ptree}{Phylogenetic tree}

\item{w.shape}{Shape parameter of the Gamma probability density function representing the generation time}

\item{w.scale}{Scale parameter of the Gamma probability density function representing the generation time}

\item{ws.shape}{Shape parameter of the Gamma probability density function representing the sampling time}

\item{ws.scale}{Scale parameter of the Gamma probability density function representing the sampling time}

\item{w.mu}{Mean of the Gamma probability density function representing the generation time, used with w.sigma to replace the shape and scale parameters}

\item{w.sigma}{Standard deviation of the Gamma probability density function representing the generation time}

\item{ws.mu}{Mean of the Gamma probability density function representing the sampling time, used with ws.sigma to replace the shape and scale parameters}

\item{ws.sigma}{Standard deviation of the Gamma probability density function representing the sampling time}

\item{mcmcIterations}{Number of MCMC iterations to run the algorithm for}

\item{thinning}{MCMC thinning interval between two sampled iterations}

\item{startNeg}{Starting value of within-host coalescent parameter Ne*g}

\item{R0}{Starting value of R0, one of startOff.r or startOff.p must also be specified}

\item{startOff.r}{Starting value of parameter off.r}

\item{startOff.p}{Starting value of parameter off.p}

\item{startPi}{Starting value of sampling proportion pi}

\item{updateNeg}{Whether of not to update the parameter Ne*g}

\item{updateOff.r}{Whether or not to update the parameter off.r}

\item{updateOff.p}{Whether or not to update the parameter off.p}

\item{updatePi}{Whether or not to update the parameter pi}

\item{startCTree}{Optional combined tree to start from}

\item{updateTTree}{Whether or not to update the transmission tree}

\item{optiStart}{Whether or not to optimise the MCMC start point}

\item{dateT}{Date when process stops (this can be Inf for fully simulated outbreaks)}

\item{epiData}{A list containing any of the following named elements: \cr \cr 
\strong{exposure} a data.frame with row.names matching the case names and columns: 
\describe{
  \item{\dQuote{start}}{the start time of the probable exposure for this case}
  \item{\dQuote{end}}{the end time of the probable exposure for this case}
} 
\strong{contact} a data.frame with columns: 
\describe{
  \item{\dQuote{case.A}}{the name of the first case in this probable transmission pair}
  \item{\dQuote{case.B}}{the name of the second case in this probable transmission pair}
  \item{\dQuote{start}}{the start of the probable contact time for the pair of cases}
  \item{\dQuote{end}}{the end of the probable contact time for the pair of cases}
}  
\strong{location} a named list of character vectors with vector names matching the case names and each vector listing the locations of that case  \cr}

\item{penalize}{whether to penalize transmission trees probability based on the \strong{epiData}}

\item{trackPenalty}{whether to save information about the penalty amounts as well as what part of the tree caused each penalty}

\item{exposureP}{numeric amounts for a single penalty in each area of the epidemiological data}

\item{contactP}{numeric amounts for a single penalty in each area of the epidemiological data}

\item{locationP}{numeric amounts for a single penalty in each area of the epidemiological data}
}
\value{
posterior sample set of transmission trees
}
\description{
Infer transmission tree given a phylogenetic tree
}
