\name{makeFullTreeFromPTree}
\alias{makeFullTreeFromPTree}
\title{Deterministically create a transmission tree compatible with the provided phylogenetic tree}
\usage{
makeFullTreeFromPTree(tree)
}
\arguments{
  \item{tree}{phylogenetic tree}
}
\value{
A non-zero probability phylogenetic+transmission tree
}
\description{
Deterministically create a transmission tree compatible
with the provided phylogenetic tree
}

