% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medTTree.R
\name{medTTree}
\alias{medTTree}
\title{Return the medoid from a MCMC output}
\usage{
medTTree(record, burnin = 0.5)
}
\arguments{
\item{record}{Output from inferTTree function}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}
}
\value{
The index of the medoid
}
\description{
Return the medoid from a MCMC output
}
