% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netExtract.R
\name{netExtractRecord}
\alias{netExtractRecord}
\title{Extract a node and edge list from a MCMC output}
\usage{
netExtractRecord(
  record,
  burnin = 0,
  colorMethod = c("all", "sampled", "time"),
  baseTree = c("maxPosterior", "consensus"),
  minimum = 0.2,
  wex = 1,
  indexShape = c("square", "triangle", "box", "circle", "dot", "star", "ellipse",
    "database", "text", "diamond")
)
}
\arguments{
\item{record}{MCMC output produced by inferTTree}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}

\item{colorMethod}{method to color the cases: 
\describe{
  \item{\dQuote{all}}{give each case a unique color}
  \item{\dQuote{sampled}}{give the sampled cases a unique color and color the unsampled cases grey}
  \item{\dQuote{time}}{give each case a color based on the time of infection}
}}

\item{baseTree}{The base tree to use: 
\describe{
  \item{\dQuote{maxPosterior}}{the transmission tree with the highest posterior probability}
  \item{\dQuote{consensus}}{the consensus transmission tree}
}}

\item{minimum}{Minimum probability for inclusion of a partition in the consensus, ignored if not using the consensus \strong{baseTree}}

\item{wex}{A weight expansion factor to multiply the edge weights by}

\item{indexShape}{shape for the index case}
}
\value{
a list of two lists one with the node list and one with the edge list
}
\description{
Extract a node and edge list from a MCMC output
}
\note{
The format of the node and edge list is ready to be used with the \strong{visnetwork} package
}
\author{
Matthew Gould
}
