% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloFromCtree.R
\name{phyloFromCtree}
\alias{phyloFromCtree}
\title{Converts a phylogenetic tree into an ape phylo object with node.comments identifying the host of each node}
\usage{
phyloFromCtree(ctree)
}
\arguments{
\item{ctree}{combined tree}
}
\value{
phylo object
}
\description{
Converts a phylogenetic tree into an ape phylo object with node.comments identifying the host of each node
}
\seealso{
phyloFromPtree
}
