% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phyloFromPtree.R
\name{phyloFromPtree}
\alias{phyloFromPtree}
\title{Converts a phylogenetic tree into an ape phylo object}
\usage{
phyloFromPtree(ptree)
}
\arguments{
\item{ptree}{phylogenetic tree}
}
\value{
phylo object
}
\description{
Converts a phylogenetic tree into an ape phylo object
}

