% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ctree}
\alias{plot.ctree}
\title{Plotting for ctree}
\usage{
\method{plot}{ctree}(x, ...)
}
\arguments{
\item{x}{Object of class ctree, ie a colored phylogenetic tree}

\item{...}{Additional parameters are passed on}
}
\value{
Plot of ctree
}
\description{
Plotting for ctree
}
\examples{
plot(simulateOutbreak())
}
