% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ptree}
\alias{plot.ptree}
\title{Plotting for ptree}
\usage{
\method{plot}{ptree}(x, ...)
}
\arguments{
\item{x}{Object of class ptree, ie  a phylogenetic tree}

\item{...}{Additional parameters are passed on}
}
\value{
Plot of ptree
}
\description{
Plotting for ptree
}
