% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ttree}
\alias{plot.ttree}
\title{Plotting for ttree}
\usage{
\method{plot}{ttree}(x, ...)
}
\arguments{
\item{x}{Object of class ttree, ie  a transmission tree}

\item{...}{Additional parameters are passed on}
}
\value{
Plot of ttree
}
\description{
Plotting for ttree
}
