% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCTree.R
\name{plotCTree}
\alias{plotCTree}
\title{Plot both phylogenetic and transmission trees using colors on the phylogeny}
\usage{
plotCTree(
  tree,
  maxTime = NA,
  showLabels = TRUE,
  alignLabels = FALSE,
  colUnsam = FALSE,
  trCol = "black",
  trShape = 20,
  cex = 1
)
}
\arguments{
\item{tree}{Combined phylogenetic/transmission tree}

\item{maxTime}{Maximum time to show on the x axis}

\item{showLabels}{Whether or not to show the labels}

\item{alignLabels}{Whether to align the labels of the tree using dotted lines from the tips}

\item{colUnsam}{Give the unsampled cases colors (TRUE), or have all unsampled cases be grey (FALSE)}

\item{trCol}{The color for the symbol annotating transmission events on the tree}

\item{trShape}{Integer code of the plotting symbol for annotating transmission events on the tree}

\item{cex}{Numeric character expansion factor for the tip labels}
}
\description{
Plot both phylogenetic and transmission trees using colors on the phylogeny
}
\examples{
plotCTree(simulateOutbreak())
}
