% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotTTree.R
\name{plotTTree}
\alias{plotTTree}
\title{Plot a transmission tree}
\usage{
plotTTree(ttree, w.shape, w.scale, showLabels = TRUE)
}
\arguments{
\item{ttree}{Transmission tree}

\item{w.shape}{Shape parameter of the Gamma probability density function representing the generation time}

\item{w.scale}{Scale parameter of the Gamma probability density function representing the generation time}

\item{showLabels}{Whether or not to show the labels}
}
\description{
Plot a transmission tree
}

