% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTTree2.R
\name{plotTTree2}
\alias{plotTTree2}
\title{Plot a transmission tree with the option of collapsing unsampled nodes}
\usage{
plotTTree2(ttree, showLabels = TRUE, collapseUnsampled = TRUE)
}
\arguments{
\item{ttree}{Transmission tree}

\item{showLabels}{Whether or not to show the labels}
}
\description{
Plot a transmission tree with the option of collapsing unsampled nodes
}

