% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ctree}
\alias{print.ctree}
\title{Print function for ctree objects}
\usage{
\method{print}{ctree}(x, ...)
}
\arguments{
\item{x}{Object of class ctree, ie a colored phylogenetic tree}

\item{...}{Additional parameters are passed on}
}
\value{
Print out details of the ctree
}
\description{
Print function for ctree objects
}
\examples{
print(simulateOutbreak())
}
