% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ptree}
\alias{print.ptree}
\title{Print function for ptree objects}
\usage{
\method{print}{ptree}(x, ...)
}
\arguments{
\item{x}{Object of class ptree, ie a phylogenetic tree}

\item{...}{Additional parameters are passed on}
}
\value{
Print out details of the ptree
}
\description{
Print function for ptree objects
}
\examples{
print(extractPTree(simulateOutbreak()))
}
