% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.resTransPhylo}
\alias{print.resTransPhylo}
\title{Print function for resTransPhylo objects}
\usage{
\method{print}{resTransPhylo}(x, ...)
}
\arguments{
\item{x}{output from inferTTree}

\item{...}{Additional parameters are passed on}
}
\value{
Print out details of TransPhylo results
}
\description{
Print function for resTransPhylo objects
}
