% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ttree}
\alias{print.ttree}
\title{Print function for ttree objects}
\usage{
\method{print}{ttree}(x, ...)
}
\arguments{
\item{x}{Object of class ttree, ie a transmission tree}

\item{...}{Additional parameters are passed on}
}
\value{
Print out details of the ttree
}
\description{
Print function for ttree objects
}
\examples{
print(extractTTree(simulateOutbreak()))
}
