% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probPTreeGivenTTree.R
\name{probPTreeGivenTTree}
\alias{probPTreeGivenTTree}
\title{Calculate the probability of a phylogenetic tree given a transmission tree}
\usage{
probPTreeGivenTTree(fulltree, neg)
}
\arguments{
\item{fulltree}{Combined phylogenetic/transmission tree}

\item{neg}{Within-host coalescent rate}
}
\value{
Probability of phylogeny given transmission tree
}
\description{
Calculate the probability of a phylogenetic tree given a transmission tree
}

