% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probPTreeGivenTTree.R
\name{probPTreeGivenTTreeR}
\alias{probPTreeGivenTTreeR}
\title{Calculate the probability of a phylogenetic tree given a transmission tree}
\usage{
probPTreeGivenTTreeR(ctree, neg, w = NULL)
}
\arguments{
\item{ctree}{Combined phylogenetic/transmission tree}

\item{neg}{Within-host coalescent rate}

\item{w}{Vector of hosts for which to calculate the probability, or NULL for all}
}
\value{
Probability of phylogeny given transmission tree
}
\description{
Calculate the probability of a phylogenetic tree given a transmission tree
}
