% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probTTree.R
\name{probTTreeR}
\alias{probTTreeR}
\title{Calculates the log-probability of a transmission tree}
\usage{
probTTreeR(ttree, off.r, off.p, pi, w.shape, w.scale, ws.shape, ws.scale,
  dateT)
}
\arguments{
\item{ttree}{Transmission tree}

\item{off.r}{First parameter of the negative binomial distribution for offspring number}

\item{off.p}{Second parameter of the negative binomial distribution for offspring number}

\item{pi}{probability of sampling an infected individual}

\item{w.shape}{Shape parameter of the Gamma probability density function representing the generation time}

\item{w.scale}{Scale parameter of the Gamma probability density function representing the generation time}

\item{ws.shape}{Shape parameter of the Gamma probability density function representing the sampling time}

\item{ws.scale}{Scale parameter of the Gamma probability density function representing the sampling time}

\item{dateT}{Date when process stops (this can be Inf for fully simulated outbreaks)}
}
\value{
Probability of the transmission tree
}
\description{
Calculates the log-probability of a transmission tree
}
