% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ptreeFromFullTree.R
\name{ptreeFromFullTree}
\alias{ptreeFromFullTree}
\title{Extracts phylogenetic tree from a combined phylogenetic/transmission tree}
\usage{
ptreeFromFullTree(tree)
}
\arguments{
\item{tree}{Combined tree}
}
\value{
phylogenetic tree
}
\description{
Extracts phylogenetic tree from a combined phylogenetic/transmission tree
}

