% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ptreeFromPhylo.R
\name{ptreeFromPhylo}
\alias{ptreeFromPhylo}
\title{Converts an ape phylo object into a phylogenetic tree}
\usage{
ptreeFromPhylo(tr, dateLastSample = 2014)
}
\arguments{
\item{tr}{phylo object}

\item{dateLastSample}{date of the last sample}
}
\value{
phylogenetic tree
}
\description{
Converts an ape phylo object into a phylogenetic tree
}

