% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseExtract.R
\name{recordCaseExtract}
\alias{recordCaseExtract}
\title{Extract various information about one case throughtout an MCMC output}
\usage{
recordCaseExtract(record, case, burnin = 0.5)
}
\arguments{
\item{record}{MCMC output produced by inferTTree}

\item{case}{String matching one of the case names in the data}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}
}
\value{
Data frame with rows over the record iterations and columns with the following data: 
\describe{
  \item{\dQuote{is_index}}{whether the case is the index case in this iteration}
  \item{\dQuote{intermediates_from_start}}{the number of intermediates in the transmission chain from this case to the index}
  \item{\dQuote{infector}}{the name of the infector of this case if sampled otherwise unsampled}
  \item{\dQuote{genTime_infector}}{the time from the infection of the infector to the time of infection of the target case}
  \item{\dQuote{time_Infected}}{time the target case was infected}
  \item{\dQuote{time_sampled}}{time the target case was sampled}
   \item{\dQuote{transmissions_to_sampled}}{the number of transmissions to sampled cases}
  \item{\dQuote{transmissions_to_unsampled}}{the number of transmissions to unsampled cases}
  \item{\dQuote{posterior}}{the posterior probability of the transmission tree at this iteration}
}
}
\description{
Extract various information about one case throughtout an MCMC output
}
\author{
Matthew Gould
}
