% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTTree.R
\name{selectTTree}
\alias{selectTTree}
\title{Select the most representative transmission tree from a MCMC output}
\usage{
selectTTree(record, burnin = 0.5)
}
\arguments{
\item{record}{Output from inferTTree function}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}
}
\value{
The index of the selected transmission tree
}
\description{
Select the most representative transmission tree from a MCMC output
}
