% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateOutbreak.R
\name{simulateContact}
\alias{simulateContact}
\title{Simulate contact data for the given ctree}
\usage{
simulateContact(ctree, min = 0.1, max = 1, propCases = 0.2)
}
\arguments{
\item{ctree}{Combined phylogenetic/transmission tree}

\item{min}{The min time difference when determining how close the contact times are to the ctree times}

\item{max}{The max time difference when determining how close the contact times are to the ctree times}

\item{propCases}{Proportion of the valid transmissions to simulate contact data for}
}
\value{
The contact.info data.frame containing the simulated contact pairs
}
\description{
Simulate contact data for the given ctree
}
\author{
Matthew Gould
}
