% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateOutbreak.R
\name{simulateExposure}
\alias{simulateExposure}
\title{Simulate exposure data for the given ctree}
\usage{
simulateExposure(ctree, min = 0.1, max = 1, propCases = 0.3)
}
\arguments{
\item{ctree}{Combined phylogenetic/transmission tree}

\item{min}{The min time difference when determining how close the exposure is to the ctree times}

\item{max}{The max time difference when determining how close the exposure is to the ctree times}

\item{propCases}{Proportion of cases to generate exporsure data for}
}
\value{
The case.info data.frame containing the simulated exposure times for each case
}
\description{
Simulate exposure data for the given ctree
}
\author{
Matthew Gould
}
