% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTTree.R
\name{startLog}
\alias{startLog}
\title{A helper function to start the log of a run
It just writes the header}
\usage{
startLog(file)
}
\arguments{
\item{file}{the log file}
}
\description{
A helper function to start the log of a run
It just writes the header
}
