% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTTree.R
\name{startNexus}
\alias{startNexus}
\title{A helper function to start the tree log of a run
It writes the taxa labels from a provided phylo object, but does not add any trees.}
\usage{
startNexus(file, tree)
}
\arguments{
\item{file}{the log file}

\item{tree}{phylo object which provides taxa labels and encodings}
}
\description{
A helper function to start the tree log of a run
It writes the taxa labels from a provided phylo object, but does not add any trees.
}
