% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.resTransPhylo}
\alias{summary.resTransPhylo}
\title{Summary function for resTransPhylo objects}
\usage{
\method{summary}{resTransPhylo}(object, ...)
}
\arguments{
\item{object}{output from inferTTree}

\item{...}{Passed on to print.phylo}
}
\value{
Print out details of TransPhylo results
}
\description{
Summary function for resTransPhylo objects
}
