% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransPhylo.R
\name{TransPhylo}
\alias{TransPhylo}
\title{Inference of Transmission Tree under a Within-Host Evolution Model}
\description{
Inference of transmission tree under a within-host evolution model.

The main functions of the package are:
\itemize{

\item inferTTree
\item simulateOutbreak
}
}
\references{
Didelot et al (2014,2017) Molecular Biology an Evolution
}
\seealso{
https://github.com/xavierdidelot/TransPhylo
}
\author{
Xavier Didelot \email{xavier.didelot@gmail.com}
}
