% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ttreeFromFullTree.R
\name{ttreeFromFullTree}
\alias{ttreeFromFullTree}
\title{Extracts transmission tree from a combined phylogenetic/transmission tree}
\usage{
ttreeFromFullTree(fulltree)
}
\arguments{
\item{fulltree}{Combined tree}
}
\value{
transmission tree
}
\description{
Extracts transmission tree from a combined phylogenetic/transmission tree
}

