% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consTTree.R
\name{consTTree}
\alias{consTTree}
\title{Build a consensus transmission tree from a MCMC output}
\usage{
consTTree(record, burnin = 0.5, minimum = 0.2, debug = F)
}
\arguments{
\item{record}{Output from inferTTree function}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}

\item{minimum}{Minimum probability for inclusion of a partition in the consensus}

\item{debug}{Used for debugging}
}
\value{
The consensus transmission tree
}
\description{
Build a consensus transmission tree from a MCMC output
}
