% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiPenTTree.R
\name{epiPenTTree}
\alias{epiPenTTree}
\title{Penalize a transmission tree based on epidemiological data}
\usage{
epiPenTTree(ttree, epiData, penaltyInfo = FALSE)
}
\arguments{
\item{ttree}{Transmission tree}

\item{epiData}{A list containing any of the following named elements: \cr \cr 
\strong{exposure} a data.frame with row.names matching the case names and columns: 
\describe{
  \item{\dQuote{start}}{the start time of the probable exposure for this case}
  \item{\dQuote{end}}{the end time of the probable exposure for this case}
} 
\strong{contact} a data.frame with columns: 
\describe{
  \item{\dQuote{case.A}}{the name of the first case in this probable transmission pair}
  \item{\dQuote{case.B}}{the name of the second case in this probable transmission pair}
  \item{\dQuote{start}}{the start of the probable contact time for the pair of cases}
  \item{\dQuote{end}}{the end of the probable contact time for the pair of cases}
}  
\strong{location} a named list of character vectors with vector names matching the case names and each vector listing the locations of that case  \cr}

\item{penaltyInfo}{Whether or not to return information about which penalties are broken for the given transmission tree}
}
\value{
Either a vector of the numeric penalties or a list of the numeric penalties and a data.frame of information about which events 
caused the penalties
}
\description{
Penalize a transmission tree based on epidemiological data
}
\author{
Matthew Gould
}
