\name{inferTTree}
\alias{inferTTree}
\title{Infer transmission tree given a phylogenetic tree}
\usage{
inferTTree(ptree, mcmcIterations = 1000)
}
\arguments{
  \item{ptree}{Phylogenetic tree}

  \item{mcmcIterations}{Number of MCMC iterations to run
  the algorithm for}
}
\value{
posterior sample set of transmission trees
}
\description{
Infer transmission tree given a phylogenetic tree
}

