% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inferTTree.R
\name{inferTTree}
\alias{inferTTree}
\title{Infer transmission tree given a phylogenetic tree}
\usage{
inferTTree(ptree, w.shape = 2, w.scale = 1, mcmcIterations = 1000,
  startNeg = 100/365, startOff.r = 1, startOff.p = 0.5, startPi = 0.5,
  updateNeg = T, updateOff.r = T, updateOff.p = T, updatePi = T,
  startFulltree = NA, updateTTree = TRUE, optiStart = T,
  datePresent = Inf)
}
\arguments{
\item{ptree}{Phylogenetic tree}

\item{w.shape}{Shape parameter of the Gamma probability density function representing the generation length w}

\item{w.scale}{Scale parameter of the Gamma probability density function representing the generation length w}

\item{mcmcIterations}{Number of MCMC iterations to run the algorithm for}

\item{startNeg}{Starting value of within-host coalescent parameter Ne*g}

\item{startOff.r}{Starting value of parameter off.r}

\item{startOff.p}{Starting value of parameter off.p}

\item{startPi}{Starting value of sampling proportion pi}

\item{updateNeg}{Whether of not to update the parameter Ne*g}

\item{updateOff.r}{Whether or not to update the parameter off.r}

\item{updateOff.p}{Whether or not to update the parameter off.p}

\item{updatePi}{Whether or not to update the parameter pi}

\item{startFulltree}{Optional combined tree to start from}

\item{updateTTree}{Whether or not to update the transmission tree}

\item{optiStart}{Whether or not to optimise the MCMC start point}

\item{datePresent}{Date when process stops (this can be Inf for fully simulated outbreaks)}
}
\value{
posterior sample set of transmission trees
}
\description{
Infer transmission tree given a phylogenetic tree
}

