% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netExtract.R
\name{netExtractTTree}
\alias{netExtractTTree}
\title{Extract a node and edge list from a transmission tree}
\usage{
netExtractTTree(
  ttree,
  colorMethod = c("all", "sampled", "time"),
  indexShape = c("square", "triangle", "dot", "star", "ellipse", "diamond")
)
}
\arguments{
\item{ttree}{}

\item{colorMethod}{method to color the cases: 
\describe{
  \item{\dQuote{all}}{give each case a unique color}
  \item{\dQuote{sampled}}{give the sampled cases a unique color and color the unsampled cases grey}
  \item{\dQuote{time}}{give each case a color based on the time of infection}
}}

\item{indexShape}{shape for the index case}
}
\value{
a list of two lists one with the node list and one with the edge list
}
\description{
Extract a node and edge list from a transmission tree
}
\note{
The format of the node and edge list is ready to be used with the \strong{visnetwork} package
}
\author{
Matthew Gould
}
