% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloFromPTree.R
\name{phyloFromPTree}
\alias{phyloFromPTree}
\title{Converts a phylogenetic tree into an ape phylo object}
\usage{
phyloFromPTree(ptree)
}
\arguments{
\item{ptree}{phylogenetic tree}
}
\value{
phylo object
}
\description{
Converts a phylogenetic tree into an ape phylo object
}
\examples{
phyloFromPTree(extractPTree(simulateOutbreak()))
}
