% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.resTransPhylo}
\alias{plot.resTransPhylo}
\title{Plotting methods}
\usage{
\method{plot}{resTransPhylo}(x, type = "tree", show.axis = T, ...)
}
\arguments{
\item{x}{Output from inferTTree}
}
\value{
Plot of TransPhylo results
}
\description{
Plotting methods
}
