% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ttree}
\alias{plot.ttree}
\title{Plotting for ttree}
\usage{
\method{plot}{ttree}(x, type = "summarised", w.shape = NA,
  w.scale = NA, ...)
}
\arguments{
\item{x}{Object of class ttree, ie  a transmission tree}

\item{type}{Type of plot to display, can be 'detailed' or 'summarised' (default)}

\item{w.shape}{Shape parameter of the generation time, needed for detailed plot only}

\item{w.scale}{Scale parameter of the generation time, needed for detailed plot only}

\item{...}{Additional parameters are passed on}
}
\value{
Plot of ttree
}
\description{
Plotting for ttree
}
\examples{
plot(extractTTree(simulateOutbreak()))
}
