\name{plotBothTree}
\alias{plotBothTree}
\title{Plot both phylogenetic and transmission trees using colors on the phylogeny}
\usage{
  plotBothTree(tree)
}
\arguments{
  \item{tree}{Combined phylogenetic/transmission tree}
}
\description{
  Plot both phylogenetic and transmission trees using
  colors on the phylogeny
}
\examples{
plotBothTree(simulateOutbreak())
}

