% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCTree.R
\name{plotCTree}
\alias{plotCTree}
\title{Plot both phylogenetic and transmission trees using colors on the phylogeny}
\usage{
plotCTree(tree, showLabels = TRUE, showStars = TRUE, cols = NA,
  maxTime = NA, cex = 1)
}
\arguments{
\item{tree}{Combined phylogenetic/transmission tree}

\item{showLabels}{Whether or not to show the labels}

\item{showStars}{Whether or not to show stars representing transmission events}

\item{cols}{Colors to use for hosts}

\item{maxTime}{Maximum time to show on the x axis}

\item{cex}{Expansion factor}
}
\value{
Returns invisibly the first parameter
}
\description{
Plot both phylogenetic and transmission trees using colors on the phylogeny
}
\examples{
plotCTree(simulateOutbreak())
}
