% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTTree.R
\name{plotTTree}
\alias{plotTTree}
\title{Plot a transmission tree in a detailed format}
\usage{
plotTTree(ttree, w.shape, w.scale, showLabels = TRUE, maxTime = NA,
  cex = 1)
}
\arguments{
\item{ttree}{Transmission tree}

\item{w.shape}{Shape parameter of the Gamma probability density function representing the generation time}

\item{w.scale}{Scale parameter of the Gamma probability density function representing the generation time}

\item{showLabels}{Whether or not to show the labels}

\item{maxTime}{Maximum value of time to show on x axis}

\item{cex}{Expansion factor}
}
\value{
Returns invisibly the first parameter
}
\description{
Plot a transmission tree in a detailed format
}
\examples{
plotTTree(extractTTree(simulateOutbreak()),2,1)
}
