% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTTree2.R
\name{plotTTree2}
\alias{plotTTree2}
\title{Plot a transmission tree in an economic format}
\usage{
plotTTree2(ttree, showLabels = TRUE, showMissingLinks = 0, cex = 1)
}
\arguments{
\item{ttree}{Transmission tree}

\item{showLabels}{Boolean for whether or not to show the labels}

\item{showMissingLinks}{Option for how to show missing links: (0) as dots, (1) as several gray levels, (2) as a single gray level}

\item{cex}{Expansion factor}
}
\value{
Returns invisibly the first parameter
}
\description{
Plot a transmission tree in an economic format
}
\examples{
plotTTree2(extractTTree(simulateOutbreak()))
}
