\name{simulateOutbreak}
\alias{simulateOutbreak}
\title{Simulate an outbreak}
\usage{
  simulateOutbreak(R, neg)
}
\arguments{
  \item{neg}{the within-host effective population size (Ne)
  times generation duration (g)}

  \item{R}{the basic reproduction number}
}
\value{
  Combined phylogenetic and transmission tree
}
\description{
  Simulate an outbreak
}

