% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulateOutbreak.R
\name{simulateOutbreak}
\alias{simulateOutbreak}
\title{Simulate an outbreak}
\usage{
simulateOutbreak(off.r = 1, off.p = 0.5, neg = 0.25, nSampled = NA,
  pi = 0.5, w.shape = 2, w.scale = 1, dateStartOutbreak = 2000,
  datePresent = Inf)
}
\arguments{
\item{off.r}{First parameter of the negative binomial distribution for offspring number}

\item{off.p}{Second parameter of the negative binomial distribution for offspring number}

\item{neg}{the within-host effective population size (Ne) timesgeneration duration (g)}

\item{nSampled}{number of sampled infected individuals, or NA for any}

\item{pi}{probability of sampling an infected individual}

\item{w.shape}{Shape parameter of the Gamma probability density function representing the generation length w}

\item{w.scale}{Scale parameter of the Gamma probability density function representing the generation length w}

\item{dateStartOutbreak}{Date when index case becomes infected}

\item{datePresent}{Date when process stops (this can be Inf for fully simulated outbreaks)}
}
\value{
Combined phylogenetic and transmission tree
}
\description{
Simulate an outbreak
}
\examples{
plotBothTree(simulateOutbreak())
}

