% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransPhylo-package.R
\docType{package}
\name{TransPhylo-package}
\alias{TransPhylo-package}
\alias{TransPhylo}
\title{Inference of Transmission Tree from a Dated Phylogeny}
\description{
Inference of transmission tree from a dated phylogeny. 
Includes methods to simulate and analyse outbreaks.
}
\references{
Didelot et al. (2014) <doi:10.1093/molbev/msu121>
Didelot et al. (2017) <doi:10.1093/molbev/msw275>.
}
\seealso{
https://github.com/xavierdidelot/TransPhylo
}
\author{
Xavier Didelot \email{xavier.didelot@gmail.com}
}
